move all item type classes under op_items.item_types
