+++ pyonepassword/api/descriptor_types.py
@@ -1,13 +1,15 @@
-from ..op_items._item_descriptor_base import OPAbstractItemDescriptor
-from ..op_items.api_credential import OPAPICredentialItemDescriptor
-from ..op_items.credit_card import OPCreditCardItemDescriptor
-from ..op_items.document import OPDocumentItemDescriptor
-from ..op_items.identity import OPIdentityItemDescriptor
-from ..op_items.login import OPLoginDescriptorItem
-from ..op_items.password import OPPasswordItemDescriptor
-from ..op_items.secure_note import OPSecureNoteItemDescriptor
-from ..op_items.server import OPServerItemDescriptor
-from ..op_items.ssh_key import OPSSHKeyItemDescriptor
+from ..op_items.item_types._item_descriptor_base import (
+    OPAbstractItemDescriptor
+)
+from ..op_items.item_types.api_credential import OPAPICredentialItemDescriptor
+from ..op_items.item_types.credit_card import OPCreditCardItemDescriptor
+from ..op_items.item_types.document import OPDocumentItemDescriptor
+from ..op_items.item_types.identity import OPIdentityItemDescriptor
+from ..op_items.item_types.login import OPLoginDescriptorItem
+from ..op_items.item_types.password import OPPasswordItemDescriptor
+from ..op_items.item_types.secure_note import OPSecureNoteItemDescriptor
+from ..op_items.item_types.server import OPServerItemDescriptor
+from ..op_items.item_types.ssh_key import OPSSHKeyItemDescriptor
 from ..op_objects import (
     OPGroupDescriptor,
     OPGroupDescriptorList,
