+++ pyonepassword/api/object_types.py
@@ -10,26 +10,32 @@ from ..op_items._new_fields import (
     OPNewUsernameField
 )
 from ..op_items._new_item import OPNewSection
-from ..op_items._op_items_base import OPAbstractItem
-from ..op_items.api_credential import OPAPICredentialItem
-from ..op_items.credit_card import OPCreditCardItem
-from ..op_items.database import OPDatabaseItem, OPDatabaseItemRelaxedValidation
-from ..op_items.document import OPDocumentFile, OPDocumentItem
 from ..op_items.item_field import OPConcealedField, OPStringField, OPTOTPField
 from ..op_items.item_field_base import OPItemField
 from ..op_items.item_section import OPSection
-from ..op_items.login import (
+from ..op_items.item_types._op_items_base import OPAbstractItem
+from ..op_items.item_types.api_credential import OPAPICredentialItem
+from ..op_items.item_types.credit_card import OPCreditCardItem
+from ..op_items.item_types.database import (
+    OPDatabaseItem,
+    OPDatabaseItemRelaxedValidation
+)
+from ..op_items.item_types.document import OPDocumentFile, OPDocumentItem
+from ..op_items.item_types.login import (
     OPLoginItem,
     OPLoginItemNewPrimaryURL,
     OPLoginItemNewURL,
     OPLoginItemRelaxedValidation,
     OPLoginItemTemplate
 )
-from ..op_items.password import OPPasswordItem
+from ..op_items.item_types.password import OPPasswordItem
+from ..op_items.item_types.secure_note import OPSecureNoteItem
+from ..op_items.item_types.server import (
+    OPServerItem,
+    OPServerItemRelaxedValidation
+)
+from ..op_items.item_types.ssh_key import OPSSHKeyItem
 from ..op_items.password_recipe import OPPasswordRecipe
-from ..op_items.secure_note import OPSecureNoteItem
-from ..op_items.server import OPServerItem, OPServerItemRelaxedValidation
-from ..op_items.ssh_key import OPSSHKeyItem
 from ..op_items.totp import OPTOTPItem
 from ..op_objects import OPGroup, OPUser, OPVault
