+++ pyonepassword/op_items/item_types/_item_descriptor_base.py
@@ -1,10 +1,10 @@
 import datetime
 from typing import List, Optional
 
-from .._abc_meta import ABCMetaDict, enforcedmethod
-from .._datetime import fromisoformat_z
-from ..json import safe_unjson
-from ..op_objects import OPVaultDescriptor
+from ..._abc_meta import ABCMetaDict, enforcedmethod
+from ..._datetime import fromisoformat_z
+from ...json import safe_unjson
+from ...op_objects import OPVaultDescriptor
 
 
 class OPAbstractItemDescriptor(dict, metaclass=ABCMetaDict):
