+++ pyonepassword/op_items/item_types/credit_card.py
@@ -1,9 +1,9 @@
+from .._item_descriptor_registry import op_register_item_descriptor_type
+from .._op_item_type_registry import op_register_item_type
+from ..item_field_base import OPItemField
+from ..item_section import OPSection
 from ._item_descriptor_base import OPAbstractItemDescriptor
-from ._item_descriptor_registry import op_register_item_descriptor_type
-from ._op_item_type_registry import op_register_item_type
 from ._op_items_base import OPAbstractItem
-from .item_field_base import OPItemField
-from .item_section import OPSection
 
 
 @op_register_item_descriptor_type
