+++ pyonepassword/op_items/item_types/login.py
@@ -1,13 +1,13 @@
 from typing import List, Optional, Union
 
+from .._item_descriptor_registry import op_register_item_descriptor_type
+from .._new_fields import OPNewPasswordField, OPNewUsernameField
+from .._new_item import OPNewItemMixin
+from .._op_item_type_registry import op_register_item_type
+from ..item_field_base import OPItemField
+from ..item_section import OPSection
 from ._item_descriptor_base import OPAbstractItemDescriptor
-from ._item_descriptor_registry import op_register_item_descriptor_type
-from ._new_fields import OPNewPasswordField, OPNewUsernameField
-from ._new_item import OPNewItemMixin
-from ._op_item_type_registry import op_register_item_type
 from ._op_items_base import OPAbstractItem
-from .item_field_base import OPItemField
-from .item_section import OPSection
 
 
 class OPNewLoginItemURLException(Exception):
