+++ pyonepassword/op_items/item_types/password.py
@@ -1,6 +1,6 @@
+from .._item_descriptor_registry import op_register_item_descriptor_type
+from .._op_item_type_registry import op_register_item_type
 from ._item_descriptor_base import OPAbstractItemDescriptor
-from ._item_descriptor_registry import op_register_item_descriptor_type
-from ._op_item_type_registry import op_register_item_type
 from ._op_items_base import OPAbstractItem
 
 # {
